Adjusted where the banner lies.
With a bit of tweaking, it looks like this is what I want it to do:
Scale to the height of the widget, maintain aspect ratio, and position
the banner on the left (well, the start; interestingly, this will likely
at least look decent in RTL locales).  The fitStart scaleType was the
kicker.