Further service fleshing.
I'm somewhat baffled as to why the decision was made to make class
statics look like a sort of object in Kotlin, but I'll admit it helps
keep them organized rather than potentially scattered hither and yon
throughout the class.

Regardless, that's the main thing added to the service.  Plus, it's now
a JobIntentService, which is likely the best idea.  It'll just scream a
BroadcastIntent into the void whenever it gets data.