Created ConnectionStateAwareApplication.
Owing to how Android Q deprecated the ability to synchronously check the
network state, I needed a different solution to how to tell if the
network's up or not (so that it can report a different error if need
be).  Enter ConnectionStateAwareApplication, a subclass of Application
that sets up connectivity observing and sends out LiveData stuff to
anyone interested, so long as that interest is in the range of when
networks come and go from the system.

I sure hope this works as planned, as subclassing Application for
something like this seems like a Genuinely Wrong™ way to do it.