Updated some widget info.
First off, this version of the widget will only resize horizontally.
Also to that end, the min width is now 180dp, which forces it to a 3x1
size.

Second, I took out the configuration activity line for now.  I'm
starting to test, and that's causing problems with it not really
existing and all.

Third, there's a new preview image.  This will likely change when I'm
closer to release, but this at least looks sort of like the actual
widget.