Tweaked some layout bits.
This at least looked right in the emulator and on my phone.  I know,
that doesn't mean much in the Android world, and I'm certain I'll get
any number of "experts" who'll claim it's "trivial" to set up an
activity to configure literally every font size used in the widget, but
we'll deal with that when we get there.  By which I mean we'll tweak the
font sizes, not that we'll regurgitate a screen full of font size
inputs.