Moved alarm and fetchery to onUpdate.
During development, I learned that when overwriting the APK when a
widget is active, onUpdate gets called but onEnabled doesn't.  Since
onUpdate is also called when a widget is added regardless (and I don't
use the normal update mechanism to get the data in), it makes more sense
to do the alarm scheduling and initial service request there.