Changed initial error string.
Turns out that, on a real device boot, it might be take a while before
the initial data comes in for the widget, depending on how long the rest
of the system takes to warm up.  So, maybe it makes more sense to use a
message that's a bit more descriptive than just "?".