Made the definition of "November" consistent.
Technically, November is still November.  However, for these purposes,
"November" also means "needs to fetch every minute".  To that end,
"November" is also when there is no data to display, either because it's
the first run or there was an error that prevented a fetch before.

Also, rearranged code such that the POSITIVE if-branch (that is, if it
IS November) comes before the NEGATIVE branch (if it is NOT November).

By the time I'm done with this, the word "November" will no longer have
any meaning to me.