Made WidgetProvider less November-dependent.
On retrospect, it turns out I don't really need to know if it's November
or not.  All I need to know is if the most recently-known run has ended.
When a new run has been scheduled, we can just treat it as a future run
and start updating every minute again, and that gets reported in the
JSON data.  To that end, WidgetProvider is now November-free, as it just
uses totalHours and runStartTimeMillis to determine if it needs to do
per-minute updates or report a future or past run.

DataFetchService, though, still uses November to avoid an Omega Shift
check.