Initial commit of the configurator and Rustproof Bee Shed option.
It'll set the setting, per se, but it's sort of ugly right now.  More to
follow.

Yes, this does make renderWidgets and renderWidget static (companion)
methods in WidgetProvider.  Turns out it was easy to do so.  Who knew?
That was needed to let the configurator call to update the widget it's
creating.