Cleaned up the layout of the configurator a bit.
First, now Rustproof Bee Shed is a checkbox.  My phone was having a very
hard time keeping its story straight as to whether or not it was night
mode; the switch thought it was night and the rest of the interface
thought it was day, meaning the switch was a bright white against the
also-white background, which wasn't working at all.  Checkboxes are
much better-colored for such confusion.

Second, the pref icon is now a drawableEnd, not its own ImageView.  Not
sure why I forgot that was a thing.  Sure saves me a lot of layout
headaches.

Third, all right and left references are now start and end in the
configurator layout.  If someone wants to translate this, that should
help immensely.

Fourth, the pref icon now updates correctly if the screen is rotated and
the option was checked.