Turn on vertical resizing for the widget.
I turned this off because I didn't think it was relevant, but it turns
out there's cases where it very well is, so on it goes.  Oops.

This may cause some visual issues if the widget is made significantly
taller than it is wide, as the banner image won't render properly.  More
research is needed.